# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Gestion - Documentation",
    "version": "1.0.0",
    "description": "Documentation complete de l'API de gestion",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Serveur de developpement"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Shop": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID unique du shop",
            "example": "66b2f0d8cdbd8f1a9d2f1c44"
          },
          "name": {
            "type": "string",
            "description": "Nom du shop",
            "example": "Boutique Centrale"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "description": "Statut du shop",
            "example": "ACTIVE"
          },
          "ownerUserId": {
            "type": "string",
            "description": "ID du proprietaire",
            "example": "66b2f0d8cdbd8f1a9d2f1c11"
          },
          "boxId": {
            "type": "string",
            "description": "ID de la box assignee",
            "nullable": true,
            "example": "66b2f0d8cdbd8f1a9d2f1c33"
          },
          "assignedBox": {
            "nullable": true,
            "$ref": "#/components/schemas/Box",
            "description": "Details de la box assignee au shop"
          },
          "activeRent": {
            "nullable": true,
            "$ref": "#/components/schemas/Rent",
            "description": "Location active associee au shop"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date de creation"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date de modification"
          }
        }
      },
      "ShopWithOwner": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID unique du shop",
            "example": "66b2f0d8cdbd8f1a9d2f1c44"
          },
          "name": {
            "type": "string",
            "description": "Nom du shop",
            "example": "Boutique Centrale"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "description": "Statut du shop",
            "example": "ACTIVE"
          },
          "ownerUserId": {
            "type": "string",
            "description": "ID du proprietaire",
            "example": "66b2f0d8cdbd8f1a9d2f1c11"
          },
          "ownerUser": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "description": "ID unique de l'utilisateur"
              },
              "profile": {
                "$ref": "#/components/schemas/UserProfile"
              }
            }
          },
          "boxId": {
            "type": "string",
            "description": "ID de la box assignee",
            "nullable": true,
            "example": "66b2f0d8cdbd8f1a9d2f1c33"
          },
          "assignedBox": {
            "nullable": true,
            "$ref": "#/components/schemas/Box",
            "description": "Details de la box assignee au shop"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date de creation"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date de modification"
          }
        }
      },
      "ShopCreate": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Nom du shop",
            "example": "Boutique Centrale"
          },
          "ownerUserId": {
            "type": "string",
            "description": "ID du proprietaire",
            "example": "66b2f0d8cdbd8f1a9d2f1c11"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "description": "Statut du shop",
            "example": "ACTIVE"
          }
        }
      },
      "ShopUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nom du shop",
            "example": "Boutique Centrale"
          },
          "ownerUserId": {
            "type": "string",
            "description": "ID du proprietaire",
            "example": "66b2f0d8cdbd8f1a9d2f1c11"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "description": "Statut du shop",
            "example": "INACTIVE"
          }
        }
      },
      "Box": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID unique de la box",
            "example": "66b2f0d8cdbd8f1a9d2f1c33"
          },
          "label": {
            "type": "string",
            "description": "Libelle de la box",
            "example": "Box A1"
          },
          "state": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "RENTED",
              "REPAIR"
            ],
            "description": "Etat de la box",
            "example": "AVAILABLE"
          },
          "rent": {
            "type": "number",
            "description": "Prix de location",
            "example": 1500
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date de creation"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date de modification"
          }
        }
      },
...
      "BoxAssignate": {
        "type": "object",
        "required": [
          "boxId",
          "isAssignate"
        ],
        "properties": {
          "rent": {
            "type": "number",
            "minimum": 0,
            "description": "Montant du loyer (obligatoire pour une assignation, ignore pour une desassignation)",
            "example": 1500
          },
          "frequency": {
            "type": "string",
            "enum": [
              "WEEKLY",
              "MONTHLY",
              "YEARLY"
            ],
            "description": "Frequence de paiement du loyer (obligatoire pour une assignation, ignore pour une desassignation)",
            "example": "MONTHLY"
          },
          "boxId": {
            "type": "string",
            "description": "ID de la box",
            "example": "66b2f0d8cdbd8f1a9d2f1c33"
          },
          "shopId": {
            "type": "string",
            "description": "ID du shop a assigner",
            "example": "66b2f0d8cdbd8f1a9d2f1c44"
          },
          "isAssignate": {
            "type": "boolean",
            "description": "true pour assigner, false pour desassigner",
            "example": true
          }
        }
      },
...
      "UserProfile": {
        "type": "object",
        "required": [
          "fullName",
          "tel"
        ],
        "properties": {
          "fullName": {
            "type": "string",
            "description": "Nom complet"
          },
          "tel": {
            "type": "string",
            "description": "Numero de telephone (10 chiffres)",
            "example": "0612345678"
          },
          "solde": {
            "type": "number",
            "description": "Solde du compte",
            "example": 0
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Adresse email"
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 10
          },
          "total": {
            "type": "integer",
            "example": 100
          },
          "pages": {
            "type": "integer",
            "example": 10
          }
        }
      },
...
      "ApiResponseMessage": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ApiResponseShop": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Shop"
          }
        }
      },
...
      "ApiResponseShopListPaged": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Shop"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
...
      "ApiResponseShopListWithOwnerPaged": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShopWithOwner"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
      "ApiResponseShopWithOwner": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ShopWithOwner"
              }
            ],
            "nullable": true
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Requete invalide",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Non authentifie",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Acces refuse",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Ressource non trouvee",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Erreur interne du serveur",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/shops": {
      "post": {
        "summary": "Creer un shop",
        "tags": [
          "Shops"
        ],
        "description": "Le ownerUserId est derive du token JWT de l'utilisateur connecte.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Le champ ownerUserId, s'il est fourni, est ignore et remplace par l'identite du token.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShopCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Shop cree avec succes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseShop"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "get": {
        "summary": "Recuperer tous les shops",
        "tags": [
          "Shops"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "Numero de page"
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 10
            },
            "description": "Nombre d'elements par page"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des shops",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseShopListPaged"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/shops/assignate": {
      "patch": {
        "summary": "Assigner ou desassigner une box a un shop",
        "tags": [
          "Shops"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoxAssignate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assignation effectuee avec succes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMessage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/shops/search": {
      "get": {
        "summary": "Rechercher des shops par criteres",
        "tags": [
          "Shops"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "INACTIVE"
              ]
            },
            "description": "Filtre par statut"
          },
          {
            "in": "query",
            "name": "ownerFullName",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filtre par nom complet du proprietaire (partiel, insensible a la casse)"
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "Numero de page"
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 10
            },
            "description": "Nombre d'elements par page"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des shops avec le profil du proprietaire",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseShopListWithOwnerPaged"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/shops/my-shop": {
      "get": {
        "summary": "Recuperer la boutique de l'utilisateur connecte",
        "tags": [
          "Shops"
        ],
        "description": "Retourne uniquement l'id et le nom de la boutique appartenant à l'utilisateur authentifie",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Informations de la boutique de l'utilisateur",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "6583d8f3e4b0a1234567890a"
                        },
                        "name": {
                          "type": "string",
                          "example": "Ma Boutique"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Aucune boutique trouvee pour cet utilisateur",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "No shop found for this user"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/shops/{id}": {
      "get": {
        "summary": "Recuperer un shop par ID avec details de box assignee",
        "tags": [
          "Shops"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID unique du shop"
          }
        ],
        "responses": {
          "200": {
            "description": "Shop trouve avec boxId et assignedBox separes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Shop"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "summary": "Mettre a jour un shop",
        "tags": [
          "Shops"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID unique du shop"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShopUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Shop mis a jour avec succes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseShop"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "summary": "Supprimer un shop",
        "tags": [
          "Shops"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID unique du shop"
          }
        ],
        "responses": {
          "200": {
            "description": "Shop supprime avec succes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMessage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/shops/owner/{ownerUserId}": {
      "get": {
        "summary": "Recuperer un shop par proprietaire",
        "tags": [
          "Shops"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "ownerUserId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID du proprietaire"
          }
        ],
        "responses": {
          "200": {
            "description": "Shop retourne (ou null si aucun) avec box assignee",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseShopWithOwner"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
...
}